+++ src/tetris/movedown.h
@@ -1,11 +1,11 @@
 bool MoveDown() { // returns true if the game has ended as a result of this move
   int y, x;
   if(PrepForMove(&y, &x)){
-    if(!curpiece_->get_plane()->move(y + 1, x)){
+    if(!curpiece_->move(y + 1, x)){
       throw TetrisNotcursesErr("move()");
     }
     if(InvalidMove()){
-      if(!curpiece_->get_plane()->move(y, x)){
+      if(!curpiece_->move(y, x)){
         throw TetrisNotcursesErr("move()");
       }
       if(y <= board_top_y_ - 1){
