+++ src/tetris/movelateral.h
@@ -2,11 +2,11 @@ void MoveLateral(int direction) { // pass in -1 for left, 1 for right
   int shift = 2 * direction;
   int y, x;
   if(PrepForMove(&y, &x)){
-    if(!curpiece_->get_plane()->move(y, x + shift)){
+    if(!curpiece_->move(y, x + shift)){
       throw TetrisNotcursesErr("move()");
     }
     if(InvalidMove()){
-      if(!curpiece_->get_plane()->move(y, x)){
+      if(!curpiece_->move(y, x)){
         throw TetrisNotcursesErr("move()");
       }
     }else{
