+++ src/tetris/stuck.h
@@ -1,19 +1,19 @@
 bool InvalidMove() { // a bit wasteful, but piece are tiny
   int dy, dx;
-  curpiece_->get_plane()->get_dim(&dy, &dx);
+  curpiece_->get_dim(&dy, &dx);
   while(dy--){
     int x = dx;
     while(x--){
       ncpp::Cell c, b;
-      if(curpiece_->get_plane()->get_at(dy, x, &c) < 0){
+      if(curpiece_->get_at(dy, x, &c) < 0){
         throw TetrisNotcursesErr("get_at()");
       }
       if(c.is_simple()){
         continue;
       }
-      curpiece_->get_plane()->release(c);
+      curpiece_->release(c);
       int transy = dy, transx = x; // need game area coordinates via translation
-      curpiece_->get_plane()->translate(*board_, &transy, &transx);
+      curpiece_->translate(*board_, &transy, &transx);
       if(transy < 0 || transy >= board_->get_dim_y() || transx < 0 || transx >= board_->get_dim_x()){
         return true;
       }
